requirements.txt
yfinance==0.2.32
pandas==2.0.3
pandas-market-calendars==4.3.1
numpy>=1.23.0
scipy>=1.9.0
fredapi==0.5.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
openpyxl==3.1.5
pytz==2023.3
pytest==7.4.0
finnhub-python==2.4.19
ib-insync==0.9.86